requirements.txt
ultralytics>=8.0.0
--extra-index-url https://download.pytorch.org/whl/cu128
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
numpy>=1.22.0
Pillow>=9.0.0
opencv-python>=4.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.2.2
gradio>=3.32.0
pandas>=2.0.0
seaborn>=0.12.2
wandb>=0.19.11
tensorboard>=2.19.0